Added "with*" methods for setting inputs
This is similar to how the value objects in the current PSR-7 draft
work. We allow for setting it all at once in the constructor if
anyone is worried about the very small overhead from creating multiple
objects (or otherwise has a reason to construct the array ahead of
time). Each object remains immutable.